namei: obey trailing magic-link DAC permissions
/* Background. */
The ability for userspace to "re-open" file descriptors through
/proc/self/fd has been a very useful tool for all sorts of usecases
(container runtimes are just one common example). However, the current
interface for doing this has resulted in some pretty subtle security
holes. Userspace can re-open a file descriptor with more permissions
than the original, which can result in cases such as /proc/$pid/exe
being re-opened O_RDWR at a later date even though (by definition)
/proc/$pid/exe cannot be opened for writing. When combined with O_PATH
the results can get even more confusing.

We cannot block this outright. Aside from userspace already depending on
it, it's a useful feature which can actually increase the security of
userspace. For instance, LXC keeps an O_PATH of the container's
/dev/pts/ptmx that gets re-opened to create new ptys and then uses
TIOCGPTPEER to get the slave end[1]. This allows for pty allocation
without resolving paths inside an (untrusted) container's rootfs. There
isn't a trivial way of doing this that is as straight-forward and safe
as O_PATH re-opening.

/* Semantics. */
Instead we have to restrict it in such a way that it doesn't break
(good) users but does block potential attackers. The solution applied in
this patch is to restrict *re-opening* (not resolution through)
magic-links by requiring that mode of the link be obeyed. Normal
symlinks have modes of a+rwx but magic-links can have other modes. These
magic-link modes were historically ignored during path resolution, but
they've now been re-purposed for more useful ends.

It is also necessary to define semantics for the mode of an O_PATH
descriptor, since re-opening a magic-link through an O_PATH needs to be
just as restricted as the corresponding magic-link -- otherwise the
above protection can be bypassed. There are two distinct cases:

 1. The target is a regular file (not a magic-link). Userspace depends
    on being able to re-open the O_PATH of a regular file, so we must
    define the mode to be a+rwx.

 2. The target is a magic-link. In this case, we simply copy the mode of
    the magic-link. This results in an O_PATH of a magic-link
    effectively acting as a no-op in terms of how much re-opening
    privileges a process has.

CAP_DAC_OVERRIDE can be used to override all of these restrictions, but
we only permit &init_userns's capabilities to affect these semantics.
The reason for this is that there isn't a clear way to track what
user_ns is the original owner of a given O_PATH chain -- thus an
unprivileged user could create a new userns and O_PATH the file
descriptor, owning it. All signs would indicate that the user really
does have CAP_DAC_OVERRIDE over the new descriptor and the protection
would be bypassed. We thus opt for the more conservative approach.

/* Special Considerations. */
In order to make sure we only impact nd_jump_link()-style jumps with
these restrictions, we introduce a new temporary LOOKUP_MAGICLINK_JUMPED
flag which is set by nd_jump_link().

Jann Horn discovered an attack in a previous version of this patch,
where an attacker could swap a single fd between a re-openable file and
a non-reopenable one and thus re-open the non-reopenable fd (exploiting
the fact that procfs inode metadata is generated somewhat-lazily). To
fix this, we need to save the magic-link's mode during nd_jump_link() --
rather than just using nd->link_inode->i_mode. A PoC of this attack (as
well as various other attack scenarios) is included as a selftest later
in the patch series.

/* Testing. */
I have run this patch on several machines for several months. So far,
the only processes which have ever hit this case ("loadkeys" and
"kbd_mode" from the kbd package[2]) gracefully handle the permission
error and do not cause any user-visible problems. In order to give users
a heads-up, a warning is output to dmesg whenever may_open_magiclink()
refuses access. If this does end up causing problems, we can expand the
permissions on initial (non-O_PATH) open() to include extra permission
bits based on whether the user could've opened the file with other
access modes at the time.

/* Future Work. */
There is one additional way in which an attacker could re-open a file
descriptor using magic-links -- by bind-mounting the magic-link and then
opening the bind-mount. This patch does not protect against this
(because it will require more invasive changes to the mount code).
Userspace can protect against this particular attack in the meantime
through seccomp (and in fact, all mainstream container runtimes have
protections against containers doing such bind-mounts). Thus, we can
safely punt on this issue for now.

[1]: commit 54ebbfb16034 ("tty: add TIOCGPTPEER ioctl")
[2]: http://git.altlinux.org/people/legion/packages/kbd.git

Cc: Jann Horn <jannh@google.com>
Cc: Kees Cook <keescook@chromium.org>
Suggested-by: Andy Lutomirski <luto@kernel.org>
Suggested-by: Christian Brauner <christian@brauner.io>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>