open: O_EMPTYPATH: procfs-less file descriptor re-opening
/* Background. */
Userspace has made use of /proc/self/fd very liberally to allow for
descriptors to be re-opened. There are a wide variety of uses for this
feature, but it has always required constructing a pathname and could
not be done without procfs mounted. The obvious solution for this is to
extend openat(2) to have an AT_EMPTY_PATH-equivalent -- O_EMPTYPATH.

Now that descriptor re-opening has been made safe through the new
magic-link resolution restrictions, we can replicate these restrictions
for O_EMPTYPATH (we compute the effective mode of the procfs magic-link,
and restrict it the same way trailing_magiclink() does).

/* Compatibility. */
Because of open(2)'s peculiar (read: complete lack of) flag checking, we
need to ensure that programs which passed garbage bits (and worked)
previously will continue to work with this change.

Passing an empty path on Linux has provided -ENOENT for an incredibly
long time, and O_EMPTYPATH acts as a no-op if the path is non-empty.
Thus it is very unlikely that any existing programs will change their
behaviour.

When doing openat(O_EMPTYPATH|O_PATH), O_PATH takes precedence and
O_EMPTYPATH is ignored. Very few users ever have a need to O_PATH
re-open an existing file descriptor, and so accommodating them at the
expense of further complicating O_PATH makes little sense. Ultimately,
if users ask for this we can always add support for this in openat2(2)
-- since openat2(2) will give an error if passed (O_EMPTYPATH|O_PATH).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>