procfs: switch magic-link modes to be more sane
Now that magic-link modes are obeyed for file re-opening purposes, some
of the pre-existing magic-link modes need to be adjusted to be more
semantically correct.

The most blatant example of this is /proc/self/exe, which had a mode of
a+rwx even though tautologically the file could never be opened for
writing (because it is the current->mm of a live process).

With the new O_PATH restrictions, changing the default mode of these
magic-links allows us to entirely block delayed-access attacks such as
we saw in CVE-2019-5736.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>