openat2: add open_how->reopen argument to restrict O_PATH reopening
/* Background */

With the introduction of magic-link reopening restrictions, userspace
programs may wish to make it so that file descriptors accessible by other
unprivileged processes (either through procfs or via SCM_RIGHTS) cannot
be reopened with certain modes.

This in particular would be useful for container runtimes to try to
reduce possible avenues of attack through implementation bugs (tricking
the runtime into opening its own files or not setting O_CLOEXEC). It
could also be used to try to reduce the ability for a delegated process
to perform a full (non-O_PATH) open of a file referenced by an O_PATH
descriptor it was given.

/* Future Work */

At the moment, due to how magic-link semantics work, it is only possible
to restrict the access mode when re-opening a magic-link as the trailing
component in a path. In the future it may be desirable to restrict a
magic-link from being used as a path component or as a file for
execve(2) -- effectively disabling the execute bit on magic-links even
when used as a non-trailing path component. If such a feature is
implemented, allowing an opt-in with REOPEN_NO_EXEC would be fairly
trivial.

/* Compatibility */

This feature can only be used with openat2(2) which is an extensible
syscall. Through the mechanism described in the openat2(2) man page, old
openat2(2) users will not enable any of the new restrictions , and new
openat2(2) users which do not enable any of the new restrictions will
continue to work on older kernels.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>