memfd: implement documented enforcement mechanism for vm.memfd_noexec=2
The previous implementation of vm.memfd_noexec=2 did not actually
enforce the usage of MFD_NOEXEC_SEAL, as any program that set MFD_EXEC
would not be affected by the "enforcement" mechanism. This contradicts
the explanation of the behaviour in include/linux/pid_namespace.h.

In fact, the only programs which were affected by the enforcement
mechanism were older programs that didn't specify MFD_NOEXEC_SEAL nor
MFD_EXEC. On the topic of older programs, it seems far less disruptive
to have vm.memfd_noexec=2 have the same behaviour as vm.memfd_noexec=1
(default to MFD_NOEXEC_SEAL if unspecified) to avoid breaking older
programs that didn't actually care about the exec bits -- which includes
the vast majority of programs that use memfd_create(2), thus allowing
users to be able to enable this sysctl without all older programs
needlessly breaking. Otherwise vm.memfd_noexec=2 would be unusable on
most general-purpose systems as it would require an audit of all of
userspace.

So, return -EPERM (not -EINVAL!) when MFD_EXEC is set with
vm.memfd_noexec=2 set, and make the behaviour when neither MFD_EXEC nor
MFD_NOEXEC_SEAL are set the same as vm.memfd_noexec=1 (MFD_NOEXEC_SEAL
is implied).

While we're at it, fix the warnings emitted by memfd_create() to use
pr_warn_ratelimited(). If the intention of the warning is to get
developers to switch to explicitly specifying if they want exec bits or
not, you need to warn them whenever they use it. The systemd version on
my box doesn't pass MFD_EXEC, making the warning useless for most
userspace developers because it was already emitted during boot. Commit
105ff5339f49 ("mm/memfd: add MFD_NOEXEC_SEAL and MFD_EXEC") mentions
that this was switched to pr_warn_once "as per review" but I couldn't
find the discussion anywhere, and the obvious issue (the ability for
unprivileged userspace to spam the kernel log) should be handled by
pr_warn_ratelimited.

This is a user-visible API change, but given that the implementation
didn't match the documentation, the feature is gated behind a sysctl,
and it was first released in a recent kernel version, the likelihood of
this breaking userspace is very mininal.

Cc: Christian Brauner <brauner@kernel.org>
Cc: stable@vger.kernel.org # v6.3+
Fixes: 105ff5339f49 ("mm/memfd: add MFD_NOEXEC_SEAL and MFD_EXEC")
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>