net: af_unix: sysctl to return -EROFS if unix socket is on a read-only mount
As per unix(7), the existing DAC semantics of AF_UNIX on Linux are not
mandated by POSIX and are (were?) a Linux-specific feature:

> On Linux, connecting to a stream socket object requires write
> permission on that socket; sending a datagram to a datagram
> socket likewise requires write permission on that socket.  POSIX
> does not make any statement about the effect of the permissions
> on a socket file, and on some systems (e.g., older BSDs), the
> socket permissions are ignored.

However, while we have checked for MAY_WRITE DAC permissions for a very
long time, whether the underlying filesystem is read-only was not
considered until now (most likely because doing mnt_wants_write would be
a little strange since opening a connection to a socket inode doesn't
actually cause any writes to the underlying filesystem, and so blocking
read-only remounts seems inadvisable).

The primary reason for doing this is that there are plenty of examples
of users thinking that a read-only bind-mount of a socket will stop
processes from connecting to the socke. However, because this setup is
currently used (in the container world, creating a read-only bind-mount
of /var/run/docker.sock into a container -- something which basically
gives root access to the host to the container) we cannot enable this
feature by default and instead need to gate it behind a new sysctl
(net.unix.verify_erofs) and log a warning to the kernel log by default
to alert users to this issue.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>